fix styling, lint issues, insert user management buttons for employers
