+++ src/Components/Utils/dialogs/ExportDataDialog.vue
@@ -76,7 +76,7 @@
       for (let appletId in this.user) {
         const applet = this.currentAccount.applets.find(applet => applet.id === appletId);
         const profile = this.user[appletId];
-        const MRN = profile.MRN.match(/None \(.*\)/) ? null : profile.MRN;
+        const MRN = profile.MRN || null;
 
         this.profileList.push({
           ...profile,
